pyproject.toml
[tool.poetry]
name = "pygeoapi-mapml-integration"
version = "0.1.0"
description = ""
authors = ["Ricardo Garcia Silva <ricardo@kartoza.com>"]
readme = "README.md"
packages = [{include = "pygeoapi_mapml_integration", from = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
flask-cors = "^4.0.1"
jinja2 = "^3.1.4"
werkzeug = "^3.0.3"
typer = "^0.12.3"
pyproj = "^3.6.1"
pygeoapi = {git = "https://github.com/geopython/pygeoapi.git"}

[tool.poetry.scripts]
mapml-pygeoapi = "pygeoapi_mapml_integration.main:app"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
